I couldn't help trying to further optimize the ressource usage for extract_matching_loci from match.py.
Running a sample run using the human genome and ~112k loci, I estimated the peak memory usage at around 18 gb by observing htop. In addition I recorded the runtime of the different parts of the function and got the following:

Setting up: 0.0751s (0.06%)
Extracting loci: 5.4964s (4.25%)
Getting loci gc bin count: 0.8073s (0.62%)
Extracting mask: 0.0763s (0.06%)
Getting background gc: 115.4431s (89.18%)
Merging bg gc, getting bg bin count: 0.4058s (0.31%)
Matching: 0.0009s (0.00%)
Extracting loci: 0.0775s (0.06%)
Verbosity analysis: 7.0401s (5.44%)
Sorting: 0.0247s (0.02%)
Total time: 129.4472s (100.00%)

When additionally providing a bigwig file, the times were:

Setting up: 0.0771s (0.04%)
Extracting loci: 23.0924s (13.46%)
Getting loci gc bin count: 0.8186s (0.48%)
Extracting mask: 0.0796s (0.05%)
Getting background gc: 123.4100s (71.93%)
Merging bg gc, getting bg bin count: 0.1639s (0.10%)
Matching: 0.0009s (0.00%)
Extracting loci: 0.0674s (0.04%)
Verbosity analysis: 23.7682s (13.85%)
Sorting: 0.0999s (0.06%)
Total time: 171.5779s (100.00%)

Hence it is clear that extracting loci and getting background gc are the most ressource demanding tasks, both in terms of time and memory usage. Both tasks essentially amount to extracting sequences from a fasta file and calculating the percent of GC and N, and in addition the total count in a window when a bigwig file is given. Hence I created a set of utility functions that can handle this process by:

1.) Generating a list of coords, either for an input loci file or for tiling a chromosome.
and from the coords either
2a) Extract the sequence and calculate the char percentage.
or
2b) Extract the signal and calculate the sum.

This approach can then handle both use cases. It relies on the python string count method, rather than encoding all the sequences as torch tensors or numpy arrays, making it much faster and also more memory efficient, since only a single locus sequence needs to be kept in memory at a time. (I do still load entire chromosomal sequences into memory for calculating background GC and N, because it is faster than streaming, but in addition to that, only a single slice is kept in memory.)

With these changes, and some additional minor ones, the ressource usage now looks as follow:

Without a bigwig file:
Around ~2gb peak memory usage
Setting up: 0.0776s (0.94%)
Extracting loci: 3.5094s (42.33%)
Getting loci gc bin count: 0.0233s (0.28%)
Extracting mask: 0.0795s (0.96%)
Getting background gc: 4.0153s (48.43%)
Merging bg gc, getting bg bin count: 0.4111s (4.96%)
Matching: 0.0009s (0.01%)
Extracting loci: 0.0763s (0.92%)
Verbosity analysis: 0.0701s (0.85%)
Sorting: 0.0275s (0.33%)
Total time: 8.2910s (100.00%)

With a bigwig file:
Around ~5gb peak memory usage
Setting up: 0.0774s (0.19%)
Extracting loci: 16.9000s (41.37%)
Getting loci gc bin count: 0.0231s (0.06%)
Extracting mask: 0.0766s (0.19%)
Getting background gc: 9.5343s (23.34%)
Merging bg gc, getting bg bin count: 0.2759s (0.68%)
Matching: 0.0010s (0.00%)
Extracting loci: 0.0738s (0.18%)
Verbosity analysis: 13.8576s (33.93%)
Sorting: 0.0273s (0.07%)
Total time: 40.8470s (100.00%)

With the tests I've run, I've been able to reproduce the output of the current script. The only exception is the case where the bigwig file misses certain chromosomes; the current script will output a signal of all zeroes in this case, giving a total count of 0. With the functions I've implemented, I instead return nan in this case. This will result in a somewhat different threshold value for filtering background regions. I think this change is quite reasonable.

With these changes the slowest part is extracting the counts from the bigwig file. This could possibly be sped up by some parallel processing, but I didn't bother to look further into that.

Thanks for creating this amazing library and for creating this intuitive implementation of extract_matching_loci. I hope these proposed changes can serve as a useful addition, and are somewhat in line with your design philosophy.