@bovagner 2024-07-23
Yes, so the five tests fail for the following reasons:
1) test_calculate_char_perc: I changed the output type from torch tensor to numpy array in accordance with the existing docstring, and because there was no need to convert the numpy array to a torch tensor. So now the gc.dtype is numpy.float64
2) test_calculate_char_perc_raises_long: Fails because I removed the assert check for the width being greater than the sequence length. In this case the function just returns an empty array, so I thought it wasn't necessary for the function to fail because of that - say you have a very short chromosome in your list of chromosomes.
3) test_calculate_char_perc_raises_sequence: Fails because I removed the check that sequence must be a string. The function will work perfectly fine with a list or tuple of characters as input. Now, it will not work if one tries to put in a one-hot-encoded sequence, so if you prefer, I can add an explicit test that the input sequence should be a str, tuple or list.
4) test_extract_and_filter_chrom: This is quite subtle, and I'm not completely sure why it happens, but it seems to have to do with numerical precision. To explain a bit more detailed, the modifications I've proposed produces the following regions dictionary when calling _extract_and_filter_chrom("tests/data/test.fa", chrom='chr1', in_window=20, out_window=20):

{20: [8, 10], 23: [1, 2, 4, 7, 9, 11, 12], 25: [0, 3, 5, 6, 13]}

The test script expects the key 23 to be 22, but otherwise the dictionary is the same. To trace back and see where the two scripts diverge, I first observe that they produce the same gc_perc vector of

[0.5, 0.45, 0.45, 0.5, 0.45, 0.5, 0.5, 0.45, 0.4, 0.45, 0.4, 0.45, 0.45, 0.5]

when adding half the gc_bin_width we arrive at

[0.51, 0.46, 0.46, 0.51, 0.46, 0.51, 0.51, 0.46, 0.41, 0.46, 0.41, 0.46, 0.46, 0.51]

to find the gc bins we do the integer division with gc_bin_width = 0.02. We can basically do that in our heads and see that 0.46 should be converted to a bin value of 23. For some reason the current version gets 22, which is also expected by the test, but this appears to be wrong. The current implementation uses a dtype of float32, while I changed this to standard numpy precision of float64, which is likely the reason behind the discrepancy. I did a few additional tests and got:
- 0.46 // 0.02 > 23.0

- t = torch.tensor([0.45], dtype=torch.float32)
(t+0.01) // 0.02 > tensor([22.])

- t = torch.tensor([0.46], dtype=torch.float32)
t // 0.02 > tensor([23.])

- t = torch.tensor([0.45], dtype=torch.float64)
(t+0.01) // 0.02 > tensor([23.], dtype=torch.float64)

- t = torch.tensor([0.46], dtype=torch.float64)
t // 0.02 > tensor([23.], dtype=torch.float64)

- t = numpy.array([0.45], dtype=numpy.float32)
(t+0.01) // 0.02 > array([22.], dtype=float32)

- t = numpy.array([0.46], dtype=numpy.float32)
t // 0.02 > array([23.], dtype=float32)

- t = numpy.array([0.45], dtype=numpy.float64)
(t+0.01) // 0.02 > array([23.])

- t = numpy.array([0.46], dtype=numpy.float64)
t // 0.02 > array([23.])

The tldr is that I think the new version I've proposed has the 'correct' output, but that in practice, it doesn't really matter, as these issues are likely to only arise with small and clean window sizes like 20. I was able to get the exact same matching loci with both implementations using a standard in_window of 2114 for ~112k input loci.

5) test_extract_matching_loci: Here the test expects 5 matching loci but only 3 are given as output. Here I assumed that the loci should be filtered based on the provided chroms, but I can see that you do not do that in your script, you only filter loci based on if they are within bounds on the chromosome. So if you want the chroms parameter to be a parameter that only determines which chroms the matching loci are found on, and not which chroms the input loci can be on, then I will make a modification that doesn't filter the input loci based on the chroms parameter.