@bovagner 2024-08-01
No problem. I updated the fork, so that peak loci are no longer filtered based on the input chroms. Hence test 5) no longer fails.

I looked a bit further into test 4) above. It is indeed related to floating point precision, and the fact that numbers like 0.01 cannot be represented exactly as floating points in a binary representation. Hence you cannot be guaranteed to get correct results even by choosing a larger floating point precision, so float64 is not necessarily better than float32. As far as I can tell, it's possible to 'round away' the slight error introduced from the addition and division in (loci_gc + gc_bin_width / 2.) // gc_bin_width by rounding before casting to an integer. So doing something like this works in the cases I have looked at:

gc_perc = ((gc_perc + gc_bin_width / 2) / gc_bin_width).round(10).astype(int) # note: normal division, not floor division

One could also just bin the gc percentages directly. For instance creating some bins:

bins = numpy.linspace(0,1,int(1/gc_bin_width)+1) 

which creates evenly sized bins, though the width are in some cases not exactly equal to gc_bin_width. Or

bins = numpy.arange(0, 1+gc_bin_width, gc_bin_width)

which creates bins with width gc_bin_width, though the last bin can extend beyond 1. The linspace approach is more numerically stable according to the docs. Both binning strategies creates one less bin than the current approach, and shifted a bit, so if you want the same binning as currently, you could add an extra bin and shift them by gc_bin_width/2, though then you may start to run into some of the same issues.

The do the binning:

gc_perc = numpy.digitize(gc_perc, bins)-1
gc_bin_count = numpy.bincount(gc_perc, minlength=len(bins)-1)

-------
The three other tests fail due to small changes in dtypes are input verification. While I think these small changes are reasonable and slight improvements, I realize you may have reasons for keeping your current checks and output dtypes, such as for reproducibility and potential future functions you would want to fail and/or work with certain inputs.