@jmschrei 2024-08-02
Thanks for being so thorough in your investigation. I'm ready to merge this but would you mind (1) updating `docs/whats_new.rst` to briefly describe your contribution and (2) updating the author list on `match.py` to include your name and contact info, if you'd like to be recognized for writing a significant portion of the code.