@bovagner 2024-08-03
Should be updated now. I've had much use of your libraries in my work, so only happy to be able to contribute.