@ghuls 2024-08-06
> I looked a bit further into test 4) above. It is indeed related to floating point precision, and the fact that numbers like 0.01 cannot be represented exactly as floating points in a binary representation. Hence you cannot be guaranteed to get correct results even by choosing a larger floating point precision, so float64 is not necessarily better than float32. As far as I can tell, it's possible to 'round away' the slight error introduced from the addition and division in (loci_gc + gc_bin_width / 2.) // gc_bin_width by rounding before casting to an integer. So doing something like this works in the cases I have looked at:

You can use `assertAlmostEqual`:
https://docs.python.org/3/library/unittest.html#unittest.TestCase.assertAlmostEqual