@ghuls 2024-08-09
It might be worth to consider using `bigtools` `average_over_bed` to get the `sum`. as it is quite a bit faster than `pybigwig`.

Semi relevant: https://github.com/jackh726/bigtools/pull/42